+++ src/main/java/com/cluttered/cryptocurrency/ann/activation/ActivationFunction.java
@@ -1,11 +1,9 @@
 package com.cluttered.cryptocurrency.ann.activation;
 
-import java.math.BigDecimal;
-
 /**
  * @author cluttered.code@gmail.com
  */
 @FunctionalInterface
 public interface ActivationFunction {
-    BigDecimal evaluate(final BigDecimal input);
+    double evaluate(final double input);
 }
